+++ openedx/core/lib/content_security_policy/README.rst
